Move authorPage components to dir
