+++ src/components/authorPage/AuthorCollection.scss
@@ -1,8 +1,8 @@
 @import '~bootstrap-sass/assets/stylesheets/bootstrap/variables';
-@import '../themes/media-queries';
-@import '../themes/common-variables';
-@import '../themes/font-mixins';
-@import '../themes/animation';
+@import '../../themes/media-queries';
+@import '../../themes/common-variables';
+@import '../../themes/font-mixins';
+@import '../../themes/animation';
 
 @-webkit-keyframes load5 {
   0%,
